Hoist AddCommand into its own class.
