+++ src/Aspire.Cli/Program.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.CommandLine;
-using System.Data;
 using System.Diagnostics;
 using System.Text;
 using Aspire.Cli.Backchannel;
@@ -67,6 +66,7 @@ public class Program
         // Commands.
         builder.Services.AddTransient<NewCommand>();
         builder.Services.AddTransient<RunCommand>();
+        builder.Services.AddTransient<AddCommand>();
 
         var app = builder.Build();
         return app;
@@ -318,203 +318,10 @@ public class Program
         parentCommand.Add(command);
     }
 
-    private static async Task<(string FriendlyName, NuGetPackage Package)> GetPackageByInteractiveFlow(IEnumerable<(string FriendlyName, NuGetPackage Package)> possiblePackages, string? preferredVersion, CancellationToken cancellationToken)
-    {
-        var distinctPackages = possiblePackages.DistinctBy(p => p.Package.Id);
-
-        var packagePrompt = new SelectionPrompt<(string FriendlyName, NuGetPackage Package)>()
-            .Title("Select an integration to add:")
-            .UseConverter(PackageNameWithFriendlyNameIfAvailable)
-            .PageSize(10)
-            .EnableSearch()
-            .HighlightStyle(Style.Parse("darkmagenta"))
-            .AddChoices(distinctPackages);
-
-        // If there is only one package, we can skip the prompt and just use it.
-        var selectedPackage = distinctPackages.Count() switch
-        {
-            1 => distinctPackages.First(),
-            > 1 => await AnsiConsole.PromptAsync(packagePrompt, cancellationToken),
-            _ => throw new InvalidOperationException("Unexpected number of packages found.")
-        };
-
-        var packageVersions = possiblePackages.Where(p => p.Package.Id == selectedPackage.Package.Id);
-
-        // If any of the package versions are an exact match for the preferred version
-        // then we can skip the version prompt and just use that version.
-        if (packageVersions.Any(p => p.Package.Version == preferredVersion))
-        {
-            var preferredVersionPackage = packageVersions.First(p => p.Package.Version == preferredVersion);
-            return preferredVersionPackage;
-        }
-
-            // ... otherwise we had better prompt.
-        var versionPrompt = new SelectionPrompt<(string FriendlyName, NuGetPackage Package)>()
-            .Title($"Select a version of {selectedPackage.Package.Id}:")
-            .UseConverter(p => p.Package.Version)
-            .EnableSearch()
-            .HighlightStyle(Style.Parse("darkmagenta"))
-            .AddChoices(packageVersions);
-
-        var version = await AnsiConsole.PromptAsync(versionPrompt, cancellationToken);
-
-        return version;
-
-        static string PackageNameWithFriendlyNameIfAvailable((string FriendlyName, NuGetPackage Package) packageWithFriendlyName)
-        {
-            if (packageWithFriendlyName.FriendlyName is { } friendlyName)
-            {
-                return $"[bold]{friendlyName}[/] ({packageWithFriendlyName.Package.Id})";
-            }
-            else
-            {
-                return packageWithFriendlyName.Package.Id;
-            }
-        }
-    }
-
-    private static (string FriendlyName, NuGetPackage Package) GenerateFriendlyName(NuGetPackage package)
-    {
-        var shortNameBuilder = new StringBuilder();
-
-        if (package.Id.StartsWith("Aspire.Hosting.Azure."))
-        {
-            shortNameBuilder.Append("az-");
-        }
-        else if (package.Id.StartsWith("Aspire.Hosting.AWS."))
-        {
-            shortNameBuilder.Append("aws-");
-        }
-        else if (package.Id.StartsWith("CommunityToolkit.Aspire.Hosting."))
-        {
-            shortNameBuilder.Append("ct-");
-        }
-
-        var lastSegment = package.Id.Split('.').Last().ToLower();
-        shortNameBuilder.Append(lastSegment);
-        return (shortNameBuilder.ToString(), package);
-    }
-
     private static void ConfigureAddCommand(Command parentCommand, IHost app)
     {
-        var command = new Command("add", "Add an integration or other resource to the Aspire project.");
-
-        var resourceArgument = new Argument<string>("resource");
-        resourceArgument.Arity = ArgumentArity.ZeroOrOne;
-        command.Arguments.Add(resourceArgument);
-
-        var projectOption = new Option<FileInfo?>("--project");
-        projectOption.Validators.Add(ProjectFileHelper.ValidateProjectOption);
-        command.Options.Add(projectOption);
-
-        var versionOption = new Option<string>("--version", "-v");
-        command.Options.Add(versionOption);
-
-        var prereleaseOption = new Option<bool>("--prerelease");
-        command.Options.Add(prereleaseOption);
-
-        var sourceOption = new Option<string?>("--source", "-s");
-        command.Options.Add(sourceOption);
-
-        command.SetAction(async (parseResult, ct) => {
-            using var activity = s_activitySource.StartActivity($"{nameof(ConfigureAddCommand)}-Action", ActivityKind.Internal);
-
-            try
-            {
-                var integrationLookup = app.Services.GetRequiredService<INuGetPackageCache>();
-
-                var integrationName = parseResult.GetValue<string>("resource");
-
-                var passedAppHostProjectFile = parseResult.GetValue<FileInfo?>("--project");
-                var effectiveAppHostProjectFile = ProjectFileHelper.UseOrFindAppHostProjectFile(passedAppHostProjectFile);
-                
-                if (effectiveAppHostProjectFile is null)
-                {
-                    return ExitCodeConstants.FailedToFindProject;
-                }
-
-                var prerelease = parseResult.GetValue<bool>("--prerelease");
-
-                var source = parseResult.GetValue<string?>("--source");
-
-                var packages = await AnsiConsole.Status().StartAsync(
-                    "Searching for Aspire packages...",
-                    context => integrationLookup.GetPackagesAsync(effectiveAppHostProjectFile, prerelease, source, ct)
-                    );
-
-                var version = parseResult.GetValue<string?>("--version");
-
-                var packagesWithShortName = packages.Select(p => GenerateFriendlyName(p));
-
-                if (!packagesWithShortName.Any())
-                {
-                    AnsiConsole.MarkupLine("[red bold]:thumbs_down: No packages found.[/]");
-                    return ExitCodeConstants.FailedToAddPackage;
-                }
-
-                var filteredPackagesWithShortName = packagesWithShortName.Where(p => p.FriendlyName == integrationName || p.Package.Id == integrationName);
-
-                if (!filteredPackagesWithShortName.Any() && integrationName is not null)
-                {
-                    // If we didn't get an exact match on the friendly name or the package ID
-                    // then try a contains search to created a broader filtered list.
-                    filteredPackagesWithShortName = packagesWithShortName.Where(
-                        p => p.FriendlyName.Contains(integrationName, StringComparison.OrdinalIgnoreCase)
-                        || p.Package.Id.Contains(integrationName, StringComparison.OrdinalIgnoreCase)
-                        );
-                }
-
-                // If we didn't match any, show a complete list. If we matched one, and its
-                // an exact match, then we still prompt, but it will only prompt for
-                // the version. If there is more than one match then we prompt.
-                var selectedNuGetPackage = filteredPackagesWithShortName.Count() switch {
-                    0 => await GetPackageByInteractiveFlow(packagesWithShortName, null, ct),
-                    1 => filteredPackagesWithShortName.First().Package.Version == version
-                        ? filteredPackagesWithShortName.First()
-                        : await GetPackageByInteractiveFlow(filteredPackagesWithShortName, null, ct),
-                    > 1 => await GetPackageByInteractiveFlow(filteredPackagesWithShortName, version, ct),
-                    _ => throw new InvalidOperationException("Unexpected number of packages found.")
-                };
-
-                var addPackageResult = await AnsiConsole.Status().StartAsync(
-                    "Adding Aspire integration...",
-                    async context => {
-                        var runner = app.Services.GetRequiredService<DotNetCliRunner>();
-                        var addPackageResult = await runner.AddPackageAsync(
-                            effectiveAppHostProjectFile,
-                            selectedNuGetPackage.Package.Id,
-                            selectedNuGetPackage.Package.Version,
-                            ct
-                            );
-
-                        return addPackageResult == 0 ? ExitCodeConstants.Success : ExitCodeConstants.FailedToAddPackage;
-                    }
-                );
-
-                if (addPackageResult != 0)
-                {
-                    AnsiConsole.MarkupLine($"[red bold]:thumbs_down: The package installation failed with exit code {addPackageResult}. For more information run with --debug switch.[/]");
-                    return ExitCodeConstants.FailedToAddPackage;
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($":thumbs_up: The package {selectedNuGetPackage.Package.Id}::{selectedNuGetPackage.Package.Version} was added successfully.");
-                    return ExitCodeConstants.Success;
-                }
-            }
-            catch (OperationCanceledException)
-            {
-                AnsiConsole.MarkupLine("[yellow bold]:stop_sign: Operation cancelled by user action.[/]");
-                return ExitCodeConstants.FailedToAddPackage;
-            }
-            catch (Exception ex)
-            {
-                AnsiConsole.MarkupLine($"[red bold]:thumbs_down: An error occurred while adding the package: {ex.Message}[/]");
-                return ExitCodeConstants.FailedToAddPackage;
-            }
-        });
-
-        parentCommand.Subcommands.Add(command);
+        var command = app.Services.GetRequiredService<AddCommand>();
+        parentCommand.Add(command);
     }
 
     public static async Task<int> Main(string[] args)
